requirements.txt
lark-oapi>=1.2.0
Flask>=2.1.3
waitress>=2.1.2
redis>=4.3.4
python-dotenv>=1.0.0
lxml>=4.9.2
Pillow>=9.4.0
requests>=2.28.1
aiohttp>=3.8.1
